+++ eth_defi/enzyme/generic_adapter_vault.py
@@ -13,6 +13,7 @@ from eth_defi.enzyme.deployment import EnzymeDeployment
 from eth_defi.enzyme.policy import create_safe_default_policy_configuration_for_generic_adapter
 from eth_defi.enzyme.vault import Vault
 from eth_defi.foundry.forge import deploy_contract_with_forge
+from eth_defi.hotwallet import HotWallet
 from eth_defi.token import TokenDetails, fetch_erc20_details
 from eth_defi.trace import assert_transaction_success_with_explanation
 from eth_defi.uniswap_v2.utils import ZERO_ADDRESS
@@ -26,7 +27,7 @@ CONTRACTS_ROOT = Path(os.path.dirname(__file__)) / ".." / ".." / "contracts"
 
 def deploy_vault_with_generic_adapter(
     deployment: EnzymeDeployment,
-    deployer: LocalAccount,
+    deployer: HotWallet,
     asset_manager: HexAddress | str,
     owner: HexAddress | str,
     usdc: Contract,
@@ -98,7 +99,7 @@ def deploy_vault_with_generic_adapter(
         Freshly deployed vault
     """
 
-    assert isinstance(deployer, LocalAccount)
+    assert isinstance(deployer, HotWallet), f"Got {type(deployer)}"
     assert asset_manager.startswith("0x")
     assert owner.startswith("0x")
 
@@ -108,6 +109,8 @@ def deploy_vault_with_generic_adapter(
     for asset in whitelisted_assets:
         assert isinstance(asset, TokenDetails)
 
+    # Log EtherScan API key
+    # Nothing bad can be done with this key, but good diagnostics is more important
     logger.info(
         "Deploying Enzyme vault. Enzyme fund deployer: %s, Terms of service: %s, USDC: %s, Etherscan API key: %s",
         deployment.contracts.fund_deployer.address,
@@ -118,22 +121,35 @@ def deploy_vault_with_generic_adapter(
 
     web3 = deployment.web3
 
-    guard = deploy_contract_with_forge(
+    guard, tx_hash = deploy_contract_with_forge(
         web3,
         CONTRACTS_ROOT / "guard",
         "GuardV0.sol",
         f"GuardV0",
-        deployer._private_key,
+        deployer,
+        etherscan_api_key=etherscan_api_key,
     )
+    logger.info("GuardV0 is %s deployed at %s", guard.address, tx_hash.hex())
     assert guard.functions.getInternalVersion().call() == 1
 
-    generic_adapter = deploy_contract(
+    # generic_adapter = deploy_contract(
+    #     web3,
+    #     f"GuardedGenericAdapter.json",
+    #     deployer,
+    #     deployment.contracts.integration_manager.address,
+    #     guard.address,
+    # )
+
+    generic_adapter, tx_hash = deploy_contract_with_forge(
         web3,
-        f"GuardedGenericAdapter.json",
+        CONTRACTS_ROOT / "in-house",
+        "GuardedGenericAdapter.sol",
+        "GuardedGenericAdapter",
         deployer,
-        deployment.contracts.integration_manager.address,
-        guard.address,
+        [deployment.contracts.integration_manager.address, guard.address],
+        etherscan_api_key=etherscan_api_key,
     )
+    logger.info("GenericAdapter is %s deployed at %s", generic_adapter.address, tx_hash.hex())
 
     policy_configuration = create_safe_default_policy_configuration_for_generic_adapter(
         deployment,
@@ -141,7 +157,7 @@ def deploy_vault_with_generic_adapter(
     )
 
     comptroller, vault = deployment.create_new_vault(
-        deployer,
+        deployer.address,
         usdc,
         policy_configuration=policy_configuration,
         fund_name=fund_name,
@@ -152,28 +168,37 @@ def deploy_vault_with_generic_adapter(
     assert vault.functions.getTrackedAssets().call() == [usdc.address]
 
     # asset manager role is the trade executor
-    vault.functions.addAssetManagers([asset_manager]).transact({"from": deployer})
+    if asset_manager != deployer.address:
+        tx_hash = vault.functions.addAssetManagers([asset_manager]).transact({"from": deployer.address})
+        assert_transaction_success_with_explanation(web3, tx_hash)
 
-    payment_forwarder = deploy_contract_with_forge(
+    # Need to resync the nonce, because it was used outside HotWallet
+    deployer.sync_nonce(web3)
+
+    payment_forwarder, tx_hash = deploy_contract_with_forge(
         web3,
-        "TermedVaultUSDCPaymentForwarder.json",
+        CONTRACTS_ROOT / "in-house",
+        "TermedVaultUSDCPaymentForwarder.sol",
+        "TermedVaultUSDCPaymentForwarder",
         deployer,
         [usdc.address, comptroller.address, terms_of_service.address],
+        etherscan_api_key=etherscan_api_key,
     )
+    logger.info("TermedVaultUSDCPaymentForwarder is %s deployed at %s", payment_forwarder.address, tx_hash.hex())
 
     # When swap is performed, the tokens will land on the integration contract
     # and this contract must be listed as the receiver.
     # Enzyme will then internally move tokens to its vault from here.
-    guard.functions.allowReceiver(generic_adapter.address, "").transact({"from": deployer})
+    guard.functions.allowReceiver(generic_adapter.address, "").transact({"from": deployer.address})
 
     # Because Enzyme does not pass the asset manager address to through integration manager,
     # we set the vault address itself as asset manager for the guard
-    tx_hash = guard.functions.allowSender(vault.address, "").transact({"from": deployer})
+    tx_hash = guard.functions.allowSender(vault.address, "").transact({"from": deployer.address})
     assert_transaction_success_with_explanation(web3, tx_hash)
 
     # Give generic adapter back reference to the vault
     assert vault.functions.getCreator().call() != ZERO_ADDRESS, f"Bad vault creator {vault.functions.getCreator().call()}"
-    tx_hash = generic_adapter.functions.bindVault(vault.address).transact({"from": deployer})
+    tx_hash = generic_adapter.functions.bindVault(vault.address).transact({"from": deployer.address})
     assert_transaction_success_with_explanation(web3, tx_hash)
 
     receipt = web3.eth.get_transaction_receipt(tx_hash)
@@ -182,20 +207,21 @@ def deploy_vault_with_generic_adapter(
     assert generic_adapter.functions.getIntegrationManager().call() == deployment.contracts.integration_manager.address
     assert comptroller.functions.getDenominationAsset().call() == usdc.address
     assert vault.functions.getTrackedAssets().call() == [usdc.address]
-    assert vault.functions.canManageAssets(asset_manager).call()
+    if asset_manager != deployer.address:
+        assert vault.functions.canManageAssets(asset_manager).call()
     assert guard.functions.isAllowedSender(vault.address).call()  # vault = asset manager for the guard
 
     usdc_token = fetch_erc20_details(web3, usdc.address)
     all_assets = [usdc_token] + whitelisted_assets
     for asset in all_assets:
         logger.info("Whitelisting %s", asset)
-        tx_hash = guard.functions.whitelistToken(asset.address, f"Whitelisting {asset.symbol}").transact({"from": deployer})
+        tx_hash = guard.functions.whitelistToken(asset.address, f"Whitelisting {asset.symbol}").transact({"from": deployer.address})
         assert_transaction_success_with_explanation(web3, tx_hash)
 
     # We cannot directly transfer the ownership to a multisig,
     # but we can set nominated ownership pending
-    if owner != deployer:
-        tx_hash = vault.functions.setNominatedOwner(owner).transact({"from": deployer})
+    if owner != deployer.address:
+        tx_hash = vault.functions.setNominatedOwner(owner).transact({"from": deployer.address})
         assert_transaction_success_with_explanation(web3, tx_hash)
         logger.info("New vault owner nominated to be %s", owner)