+++ eth_defi/foundry/forge.py
@@ -5,6 +5,7 @@
 - Verify smart contracts on Etherscan
 """
 import contextlib
+import datetime
 import io
 import logging
 import os
@@ -12,7 +13,7 @@ import os
 from pathlib import Path
 from shutil import which
 from subprocess import DEVNULL, PIPE
-from typing import Collection
+from typing import Collection, Tuple
 
 import psutil
 from eth_typing import ChecksumAddress, HexAddress, HexStr
@@ -21,7 +22,10 @@ from web3 import Web3
 from web3.contract import Contract
 
 from eth_defi.abi import get_deployed_contract
+from eth_defi.confirmation import wait_transactions_to_complete
 from eth_defi.deploy import register_contract
+from eth_defi.hotwallet import HotWallet
+from eth_defi.trace import assert_transaction_success_with_explanation
 
 logger = logging.getLogger(__name__)
 
@@ -35,18 +39,23 @@ class ForgeFailed(Exception):
     """Forge command failed."""
 
 
+#: Because of Forge's
+#:
+_last_deploy: datetime.datetime | None = None
+
+
 def _exec_cmd(
     cmd_line: list[str],
     censored_command: str,
     timeout=DEFAULT_TIMEOUT,
-) -> str:
+) -> Tuple[str, str]:
     """Execute the command line.
 
     :param timeout:
         Timeout in seconds
 
     :return:
-        Deployed contract address
+        Tuple(deployed contract address, tx hash)
     """
 
     for x in cmd_line:
@@ -64,12 +73,20 @@ def _exec_cmd(
 
     logger.debug("forge result:\n%s", output)
 
+    address = tx_hash = None
+
     for line in output.split("\n"):
         # Deployed to: 0x604Da6680Cb97A87403600B9AafBE60eeda97CA4
         if line.startswith("Deployed to: "):
-            return line.split(":")[1].strip()
+            address = line.split(":")[1].strip()
 
-    raise ForgeFailed(f"Could not parse forge output:\n{output}")
+        if line.startswith("Transaction hash: "):
+            tx_hash = line.split(":")[1].strip()
+
+    if not (address and tx_hash):
+        raise ForgeFailed(f"Could not parse forge output:\n{output}")
+
+    return address, tx_hash
 
 
 def deploy_contract_with_forge(
@@ -77,12 +94,13 @@ def deploy_contract_with_forge(
     project_folder: Path,
     contract_file: Path | str,
     contract_name: str,
-    private_key: HexBytes,
+    deployer: HotWallet,
     constructor_args: list[str] | None = None,
     etherscan_api_key: str | None = None,
     register_for_tracing=True,
     timeout=DEFAULT_TIMEOUT,
-) -> Contract:
+    wait_for_block_confirmations=0,
+) -> Tuple[Contract, HexBytes]:
     """Deploys a new contract using Forge command from Foundry.
 
     - Uses Forge to verify the contract on Etherscan
@@ -100,6 +118,11 @@ def deploy_contract_with_forge(
 
     :param web3:
         Web3 instance
+
+    :param deployer:
+        Deployer tracked as a hot wallet.
+
+        We need to be able to manually track the nonce across multiple contract deployments.
 
     :param project_folder:
         Foundry project with ``foundry.toml` in the root.
@@ -124,16 +147,22 @@ def deploy_contract_with_forge(
 
         See :py:func:`get_contract_registry`
 
-    :raise ContractDeploymentFailed:
+    :param wait_for_block_confirmations:
+        Currently not used.
+
+    :raise ForgeFailed:
         In the case we could not deploy the contract.
 
+        - Running forge failed
+        - Transaction could not be confirmed
+
     :return:
-        Contract proxy instance
+        Contract and deployment tx hash.
+
     """
     assert isinstance(project_folder, Path)
-    assert isinstance(contract_file, Path)
     assert type(contract_name) == str
-    assert isinstance(private_key, HexBytes), f"Got private key: {type(private_key)}"
+    assert isinstance(deployer, HotWallet), f"Got deployer: {type(deployer)}"
 
     if constructor_args is None:
         constructor_args = []
@@ -141,6 +170,7 @@ def deploy_contract_with_forge(
     if type(contract_file) == str:
         contract_file = Path(contract_file)
 
+    assert isinstance(contract_file, Path)
     assert type(constructor_args) in (list, tuple)
 
     json_rpc_url = web3.provider.endpoint_uri
@@ -154,6 +184,7 @@ def deploy_contract_with_forge(
         forge,
         "create",
         "--rpc-url", json_rpc_url,
+        "--nonce", str(deployer.allocate_nonce()),
     ]
 
     if etherscan_api_key:
@@ -177,7 +208,7 @@ def deploy_contract_with_forge(
 
     logger.info(
         "Deploying a contract with forge. Working directory %s, forge command: %s",
-        project_folder,
+        project_folder.resolve(),
         censored_command,
     )
 
@@ -185,7 +216,7 @@ def deploy_contract_with_forge(
     cmd_line = [
         forge,
         "create",
-        "--private-key", private_key.hex(),
+        "--private-key", deployer.private_key.hex(),
     ] + cmd_line[2:]
 
     with contextlib.chdir(project_folder):
@@ -195,7 +226,7 @@ def deploy_contract_with_forge(
         assert src_contract_file.exists(), f"Contract does not exist: {src_contract_file}, current working directory is {os.getcwd()}"
 
         # Run forge
-        contract_address = _exec_cmd(cmd_line, timeout=timeout, censored_command=censored_command)
+        contract_address, tx_hash = _exec_cmd(cmd_line, timeout=timeout, censored_command=censored_command)
 
         # Check we produced an ABI file, or was created earlier
         contract_abi = project_folder / "out" / contract_file / f"{contract_name}.json"
@@ -213,4 +244,11 @@ def deploy_contract_with_forge(
         instance.name = contract_name
         register_contract(web3, contract_address, instance)
 
-    return instance
+    tx_hash = HexBytes(tx_hash)
+    assert_transaction_success_with_explanation(
+        web3,
+        tx_hash,
+        RaisedException=ForgeFailed,
+    )
+
+    return instance, tx_hash